Convert NavigationComponent into standalone nav.js script to avoid relying on BaseComponent class.
