+++ front-end/src/index.html
@@ -4,12 +4,15 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <script type="module" src="./main.js" defer></script>
+    <script type="module" src="./nav/nav.js" defer></script>
+
     <link rel="stylesheet" href="./pages/main/stylesBase.css" />
     <title>Cherish Calendar</title>
   </head>
   <body>
     <div class="app">
       <div id="views"></div>
+      <div class="nav" style="display: flex"></div>
     </div>
   </body>
 </html>
