+++ front-end/src/main.js
@@ -1,15 +1,12 @@
 import { DayComponent } from "./pages/day/DayComponent.js";
 import { JournalComponent } from "./pages/day/journal/JournalComponent.js";
 import { CalendarComponent } from "./pages/calendar/CalendarComponent.js";
-import { NavigationComponent } from "./nav/NavigationComponent.js";
 import { EventHub } from "./eventhub/EventHub.js";
 import { Events } from "./eventhub/Events.js";
-
 
 new CalendarComponent();
 new DayComponent();
 new JournalComponent();
-new NavigationComponent();
 console.log("Everything loaded");
 
 const today = new Date();
@@ -24,4 +21,3 @@ const date = {
 };
 
 EventHub.getInstance().publish(Events.LoadMainPage, date);
-EventHub.getInstance().publish(Events.LoadNav, date);
